dev/apps/api/pyproject.toml
[project]
name = "auto-news"
version = "0.1.0"
description = "AutoNews backend"
authors = [
    { name = "Yuzhang Hu", email = "hyzwowtools@gmail.com" }
]
dependencies = [
    # --- API & Scheduler (신규) ---
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "apscheduler>=3.10,<4",
    "psycopg2-binary>=2.9",
    # --- Core ---
    "SpeechRecognition",
    "arxiv",
    "bs4",
    "chromadb>=0.4",
    "docker",
    "duckduckgo_search",
    "fake_useragent",
    "feedparser",
    "httpcore>=1.0",
    "httpx",
    "langchain==0.3.1",
    "langchain-community==0.3.1",
    "langchain-google-genai",
    "llama-index>=0.9",
    "lz4",
    "minio",
    "miniopy-async",
    "notion-client",
    "openai",
    "openai-whisper",
    "pillow",
    "pinecone-client",
    "ag2",
    "pymilvus>=2.4",
    "pymupdf",
    "python-dotenv",
    "python-snappy",
    "pyttsx3",
    "pytube",
    "pytz",
    "redis",
    "requests",
    "retry",
    "scales",
    "scylla-driver",
    "soundfile",
    "tiktoken",
    "transformers",
    "tweepy",
    "validators",
    "youtube-transcript-api",
    "yt-dlp",
    "google-generativeai>=0.8.5",
    "litellm>=1.80.10",
    "scikit-learn>=1.6.1",
]
readme = "README.md"
requires-python = ">= 3.10"

[project.scripts]
# hello = "airflow:hello"

[project.urls]
Homepage = "https://github.com/finaldie/auto-news"
Documentation = "https://github.com/finaldie/auto-news/wiki"
Repository = "https://github.com/finaldie/auto-news"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["./"]
